src/main/java/study/greedy/p42862/Solution.java
package study.greedy.p42862;

import java.util.Arrays;

class Solution {
    public int solution(int n, int[] lost, int[] reserve) {
        Arrays.sort(lost);
        Arrays.sort(reserve);

        int lostCount = n - lost.length;

        for (int i = 0; i < lost.length; i++) {
            for (int j = 0; j < reserve.length; j++) {
                if (lost[i] == reserve[j]) {
                    lost[i] = -9;
                    reserve[j] = -9;
                    lostCount++;
                    break;
                }
            }
        }

        for (int i = 0; i < lost.length; i++) {
            for (int j = 0; j < reserve.length; j++) {
                if (lost[i] == reserve[j] - 1 || lost[i] == reserve[j] + 1) {
                    lostCount++;
                    reserve[j] = -9;
                    lost[i] = -9;
                    break;
                }
            }
        }
        return lostCount;
    }
}